+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update content in `./README.md`
 - Add first publish section in `./README.md`
 - Support generating typings
-- Add `scripts["{docs,docs:watch}"]` in `./package.json`
+- Add `scripts["{docs,docs:serve,docs:watch}"]` in `./package.json`
 - Add `gh-pages` job in workflow
 - Update `./tsconfig.json`
 - Split `scripts.dev` into `./scripts/dev.sh`